'Fundations' section has been refactored + pagination has been fixed
