+++ my-app/src/components/HomeFundationsDisplay.js
@@ -1,14 +1,11 @@
 import React from "react";
 import {useState, useEffect} from "react";
-import Pagination from "./Pagination";
+import ReactPaginate from 'react-paginate';
 
 const HomeFundationsDisplay = ({data, header}) => {
 
     const [fundationsList, setFundationsList] = useState(data);
     const [currentPage, setCurrentPage] = useState(1);
-
-    console.log(data);
-    console.log(fundationsList);
 
     useEffect( ()=>{
         setFundationsList(data);
@@ -19,6 +16,11 @@ const HomeFundationsDisplay = ({data, header}) => {
     const indexOfLastFundation = currentPage * fundationsPerPage;
     const indexOfFirstFundation = indexOfLastFundation - fundationsPerPage;
     const currentFundations = fundationsList?.slice(indexOfFirstFundation, indexOfLastFundation);
+
+    const pageNumbers = [];
+    for (let i = 1; i <= Math.ceil(fundationsList.length/fundationsPerPage); i++){
+        pageNumbers.push(i);
+    }
 
     const paginate = pageNumber =>{
         setCurrentPage(pageNumber.selected+1);
@@ -37,7 +39,21 @@ const HomeFundationsDisplay = ({data, header}) => {
                                                     <p className="homeFundationsDisplay-fundationNeeds">{fundation.needs}</p>
                                                 </li>)}
             </ul>
-            <Pagination positionsTotal={fundationsList.length} positionsPerPage={fundationsPerPage} paginate={paginate}></Pagination>
+            {pageNumbers.length > 1 && 
+            <ReactPaginate
+            previousLabel={""}
+            nextLabel={""}
+            breakLabel={"..."}
+            breakClassName={"break"}
+            pageCount={pageNumbers.length}
+            forcePage={currentPage-1}
+            marginPagesDisplayed={2}
+            pageRangeDisplayed={3}
+            onPageChange={paginate}
+            containerClassName={"pagination-container"}
+            subContainerClassName={"pagination-page"}
+            activeClassName={"pagination-current-page"}
+            />}        
         </div>
     )
 }
